Add files via upload
Data was found to be extracted for 7 days versus 6.  It then allowed for a more representative report for the KPI's of the week prior.  Goal here was to create a "generic" script that can potentially be used by several site with room for expansion.  End goal is to create a script that can be run off an executable.  It shall work raw data in csv format then create an outlook email for specific DL's and allow for personal synopsis for the week.